+++ logger-logback/src/test/resources/logback-chronicle-config.xml
@@ -32,7 +32,7 @@
     <!-- ******************************************************************* -->
 
     <appender name  = "CONFIG-BINARY-CHRONICLE"
-              class = "net.openhft.chronicle.logger.logback.ChronicleAppender">
+              class = "net.openhft.chronicle.logger.logback.BlockingChronicleAppender">
         <path>${java.io.tmpdir}/chronicle-logback/config-binary-chronicle</path>
         <chronicleConfig>
             <blockSize>128</blockSize>
