+++ logger-logback/src/test/resources/logback-chronicle-zstd-binary-appender.xml
@@ -32,7 +32,7 @@
     <!-- ******************************************************************* -->
 
     <appender name  = "BINARY-ZSTD-CHRONICLE"
-              class = "net.openhft.chronicle.logger.logback.ChronicleAppender">
+              class = "net.openhft.chronicle.logger.logback.BlockingChronicleAppender">
         <path>${java.io.tmpdir}/chronicle-logback/binary-zstd-chronicle</path>
         <contentType>text/plain</contentType>
         <contentEncoding>zstd</contentEncoding>
